Modifying part 1 to work with Python 3

+++ engine.py
@@ -7,8 +7,8 @@ def main():
     screen_width = 80
     screen_height = 50
 
-    player_x = screen_width / 2
-    player_y = screen_height / 2
+    player_x = int(screen_width / 2)
+    player_y = int(screen_height / 2)
 
     libtcod.console_set_custom_font('arial10x10.png', libtcod.FONT_TYPE_GREYSCALE | libtcod.FONT_LAYOUT_TCOD)
